requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
python-gitlab==1.15.0
requests==2.22.0
six==1.13.0
urllib3==1.25.7
